requirements.txt
hydra-core==1.0.3
omegaconf==2.0.2
numpy==1.19.2
soundfile==0.10.3.post1